# services/gateway/openapi.yaml
openapi: 3.0.3
info:
  title: trust-bank API Gateway
  version: 0.1.0
  description: >
    The public API for external bank/developer integration with trust-bank —
    distinct from a tenant's own product backend (e.g. trustpay-backend),
    which serves that tenant's *own* consumer app. This is what a bank's
    engineers integrate against directly: accounts, transfers, identity
    verification, compliance checks, and card issuance/lifecycle, proxied
    to Ledger/Payments/Compliance/Cards with tiered API keys, per-key
    rate limiting, and per-backend circuit breaking in front. A
    sandbox-tier key is genuinely isolated — its own tenant in
    Ledger/Payments/Compliance, not just a lower rate limit on the same
    production data (see the /sandbox admin route below) — while every
    route below is called with the same URL
    (/v1/tenants/{realTenantId}/...) regardless of which tier issued the
    key.


    Every route below was exercised for real during this service's own
    build — not written speculatively. Response shapes are the actual
    shapes returned, captured from a live run.
  contact:
    url: https://github.com/imransaadullah/trust-bank
servers:
  - url: https://api.trustbank.example.com
    description: Production (see deploy/Caddyfile.example)
  - url: http://localhost:8084
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /v1/tenants/{tenantId}/compliance/kyc-policy:
    get:
      summary: Read the current KYC-tier policy
      description: >
        Admin-tier key required — a separate, admin-scope compliance
        credential from the operate-scope one kyc-tier-check/device-check
        use (see TenantBackendCredential's schema comment). 404 with
        code NO_POLICY_CONFIGURED if none has ever been published for
        this tier.
      tags: [Admin]
      parameters:
        - { $ref: '#/components/parameters/TenantId' }
        - { name: tier, in: query, required: true, schema: { type: integer } }
        - { name: jurisdiction, in: query, schema: { type: string, default: NG } }
      responses:
        '200':
          description: The current versioned KYCTierPolicy row
        '404':
          description: No policy published for this tier yet
    post:
      summary: Publish a new KYC-tier policy version
      description: Admin-tier key required. Publishing never updates a row — it inserts a new, higher version.
      tags: [Admin]
      parameters: [{ $ref: '#/components/parameters/TenantId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tier, requiredVerifications]
              properties:
                jurisdiction: { type: string, default: NG }
                tier: { type: integer }
                requiredVerifications: { type: array, items: { type: string } }
                dailyLimitKobo: { type: integer }
                singleTxnLimitKobo: { type: integer }
                effectiveFrom: { type: string, format: date-time }
      responses:
        '201':
          description: The newly published policy version
  /v1/tenants/{tenantId}/compliance/device-policy:
    get:
      summary: Read the current device-binding policy
      tags: [Admin]
      parameters:
        - { $ref: '#/components/parameters/TenantId' }
        - { name: jurisdiction, in: query, schema: { type: string, default: NG } }
      responses:
        '200': { description: The current versioned DeviceBindingPolicy row }
        '404': { description: No policy published yet }
    post:
      summary: Publish a new device-binding policy version
      tags: [Admin]
      parameters: [{ $ref: '#/components/parameters/TenantId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [newDeviceCooldownHours, newDeviceLimitKobo]
              properties:
                jurisdiction: { type: string, default: NG }
                maxActiveDevices: { type: integer }
                newDeviceCooldownHours: { type: integer }
                newDeviceLimitKobo: { type: integer }
                effectiveFrom: { type: string, format: date-time }
      responses:
        '201': { description: The newly published policy version }
  /v1/tenants/{tenantId}/compliance/loan-eligibility-policy:
    get:
      summary: Read the current loan-eligibility policy
      tags: [Admin]
      parameters:
        - { $ref: '#/components/parameters/TenantId' }
        - { name: jurisdiction, in: query, schema: { type: string, default: NG } }
      responses:
        '200': { description: The current versioned LoanEligibilityPolicy row }
        '404': { description: No policy published yet }
    post:
      summary: Publish a new loan-eligibility policy version
      tags: [Admin]
      parameters: [{ $ref: '#/components/parameters/TenantId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [minKycTier, maxLoanAmountKobo, maxTenorDays, interestRateAnnualBps]
              properties:
                jurisdiction: { type: string, default: NG }
                minKycTier: { type: integer }
                maxLoanAmountKobo: { type: integer }
                maxTenorDays: { type: integer }
                interestRateAnnualBps: { type: integer }
                effectiveFrom: { type: string, format: date-time }
      responses:
        '201': { description: The newly published policy version }
  /v1/tenants/{tenantId}/compliance/card-issuance-policy:
    get:
      summary: Read the current card-issuance policy
      tags: [Admin]
      parameters:
        - { $ref: '#/components/parameters/TenantId' }
        - { name: jurisdiction, in: query, schema: { type: string, default: NG } }
      responses:
        '200': { description: The current versioned CardIssuancePolicy row }
        '404': { description: No policy published yet }
    post:
      summary: Publish a new card-issuance policy version
      tags: [Admin]
      parameters: [{ $ref: '#/components/parameters/TenantId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [minKycTier, maxCardsPerCustomer, dailySpendLimitKobo, singleTxnLimitKobo]
              properties:
                jurisdiction: { type: string, default: NG }
                minKycTier: { type: integer }
                maxCardsPerCustomer: { type: integer }
                dailySpendLimitKobo: { type: integer }
                singleTxnLimitKobo: { type: integer }
                effectiveFrom: { type: string, format: date-time }
      responses:
        '201': { description: The newly published policy version }
  /v1/tenants/{tenantId}/compliance/monitoring-policy:
    get:
      summary: Read the current transaction-monitoring policy
      tags: [Admin]
      parameters:
        - { $ref: '#/components/parameters/TenantId' }
        - { name: jurisdiction, in: query, schema: { type: string, default: NG } }
      responses:
        '200': { description: The current versioned TransactionMonitoringPolicy row }
        '404': { description: No policy published yet }
    post:
      summary: Publish a new transaction-monitoring policy version
      tags: [Admin]
      parameters: [{ $ref: '#/components/parameters/TenantId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - velocityWindowHours
                - velocityMaxTransactionCount
                - velocityMaxDistinctCounterparties
                - structuringThresholdKobo
                - structuringWindowCount
                - largeSingleTxnThresholdKobo
              properties:
                jurisdiction: { type: string, default: NG }
                velocityWindowHours: { type: integer }
                velocityMaxTransactionCount: { type: integer }
                velocityMaxDistinctCounterparties: { type: integer }
                structuringThresholdKobo: { type: integer }
                structuringWindowCount: { type: integer }
                largeSingleTxnThresholdKobo: { type: integer }
                effectiveFrom: { type: string, format: date-time }
      responses:
        '201': { description: The newly published policy version }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        `gw_live_...`. Admin-tier for the Admin-tagged routes above;
        sandbox or production tier for everything else. A key only works
        for the tenant it was issued to — the tenantId in the path must
        match. A sandbox-tier key is transparently resolved to that
        tenant's isolated sandbox twin before the request ever reaches a
        backend — see the /sandbox admin route.
  parameters:
    TenantId:
      name: tenantId
      in: path
      required: true
      schema: { type: string, format: uuid }
...